requirements/base.txt
boto3
channels
channels_redis
coreapi
django
django-autocomplete-light
django-filter
django-pyodbc-azure
django-storages
django-taggit
django-taggit-serializer
djangorestframework>=3.9
markdown
pillow
psycopg2-binary
python-dateutil
requests
sentry-sdk
uwsgi
